backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mediamind-backend"
version = "0.1.0"
description = "AI-powered document and multimedia QA backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
  "sqlalchemy>=2.0,<3.0",
  "pydantic-settings>=2.6,<3.0",
  "python-multipart>=0.0.9,<1.0",
  "pypdf>=5.0,<6.0",
  "openai>=1.50,<2.0",
  "psycopg[binary]>=3.2,<4.0",
  "redis>=5.2,<6.0",
  "faiss-cpu>=1.9,<2.0",
  "transformers>=4.40,<5.0",
  "torch>=2.0,<3.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27,<1.0",
  "pytest>=8.3,<9.0",
  "pytest-cov>=5.0,<6.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing --cov-fail-under=95"
